+++ module/CSV_Political_extractor/pom.xml
@@ -2,60 +2,60 @@
 <project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
          xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<groupId>cz.cuni.mff.xrg.odcs</groupId>
-		<artifactId>module-base</artifactId>
-		<version>1.0.0</version>
-		<relativePath>../module-base/pom.xml</relativePath>
-	</parent>
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>cz.cuni.mff.xrg.odcs</groupId>
+        <artifactId>module-base</artifactId>
+        <version>1.0.0</version>
+        <relativePath>../module-base/pom.xml</relativePath>
+    </parent>
 
-	<artifactId>CSV_Political_extractor</artifactId>
-	<packaging>bundle</packaging>
-	<description>Extracts RDF data from a file.</description>
+    <artifactId>CSV_Political_extractor</artifactId>
+    <packaging>bundle</packaging>
+    <description>Extracts RDF data from a file.</description>
 
-	<properties>
-		<dpu.package>cz.cuni.mff.xrg.odcs.extractor.file</dpu.package>
-		<dpu.mainClass>CsvPoliticalExtractor</dpu.mainClass>
-		<solr.version>3.6.1</solr.version>
-	</properties>
+    <properties>
+        <dpu.package>cz.cuni.mff.xrg.odcs.extractor.file</dpu.package>
+        <dpu.mainClass>CsvPoliticalExtractor</dpu.mainClass>
+        <solr.version>3.6.1</solr.version>
+    </properties>
 
-	<dependencies>
-		<dependency>
-			<groupId>org.apache.solr</groupId>
-			<artifactId>solr-solrj</artifactId>
-			<version>${solr.version}</version>
-			
-		
-		</dependency>
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.solr</groupId>
+            <artifactId>solr-solrj</artifactId>
+            <version>${solr.version}</version>
 
-		<dependency>
-			<groupId>commons-beanutils</groupId>
-			<artifactId>commons-beanutils</artifactId>
-			<version>1.8.3</version>
-			<scope>provided</scope>
 
-		</dependency>
+        </dependency>
 
-		<dependency>
-			<groupId>org.apache.commons</groupId>
-			<artifactId>commons-lang3</artifactId>
-			<version>3.1</version>
-			<scope>provided</scope>
-		</dependency>
+        <dependency>
+            <groupId>commons-beanutils</groupId>
+            <artifactId>commons-beanutils</artifactId>
+            <version>1.8.3</version>
+            <scope>provided</scope>
 
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>4.8.2</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.codehaus.groovy</groupId>
-			<artifactId>groovy-all</artifactId>
-			<version>1.7.2</version>
-			<scope>provided</scope>
-		</dependency>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-lang3</artifactId>
+            <version>3.1</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.8.2</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.codehaus.groovy</groupId>
+            <artifactId>groovy-all</artifactId>
+            <version>1.7.2</version>
+            <scope>provided</scope>
+        </dependency>
 
 
     </dependencies>
