+++ module/CSV_Political_extractor/src/test/cz/cuni/mff/xrg/odcs/extractor/Test.java
@@ -21,7 +21,7 @@ public class Test {
             String input = null;
             env.run(extractor);
         } catch (Exception e) {
-            e.printStackTrace();  //To change body of catch statement use File | Settings | File Templates.
+            e.printStackTrace(); // To change body of catch statement use File | Settings | File Templates.
         } finally {
             // release resources
             env.release();
